pyproject.toml
[project]
name = "order-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "aiokafka>=0.13.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "fastapi>=0.135.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.2.1",
    "ruff>=0.15.5",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
